Refactor inference service code
