+++ edsl/inference_services/__init__.py
@@ -1,0 +1,5 @@
+from .inference_services_collection import InferenceServicesCollection
+from .registry import default
+from .inference_service_abc import InferenceServiceABC
+from .data_structures import AvailableModels
+from .exceptions import InferenceServiceError
