+++ edsl/inference_services/available_model_fetcher.py
@@ -2,19 +2,13 @@ from typing import Any, List, Tuple, Optional, Dict, TYPE_CHECKING, Union, Gener
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from collections import UserList
 
-from edsl.inference_services.ServiceAvailability import ServiceAvailability
+from .service_availability import ServiceAvailability
 from .inference_service_abc import InferenceServiceABC
-from edsl.inference_services.data_structures import ModelNamesList
-from edsl.enums import InferenceServiceLiteral
+from ..enums import InferenceServiceLiteral
 
-from edsl.inference_services.data_structures import LanguageModelInfo
-from edsl.inference_services.AvailableModelCacheHandler import (
-    AvailableModelCacheHandler,
-)
+from .data_structures import ModelNamesList, LanguageModelInfo, AvailableModels
 
-
-from edsl.inference_services.data_structures import AvailableModels
-
+from .available_model_cache_handler import AvailableModelCacheHandler
 
 class AvailableModelFetcher:
     """Fetches available models from the various services with JSON caching."""
