+++ edsl/inference_services/services/azure_ai.py
@@ -2,8 +2,8 @@ import os
 from typing import Any, Optional, List
 import re
 from openai import AsyncAzureOpenAI
-from .inference_service_abc import InferenceServiceABC
-from ..language_models import LanguageModel
+from ..inference_service_abc import InferenceServiceABC
+from ...language_models import LanguageModel
 
 from azure.ai.inference.aio import ChatCompletionsClient
 from azure.core.credentials import AzureKeyCredential
