+++ edsl/inference_services/services/deep_seek_service.py
@@ -3,10 +3,8 @@ import json
 import requests
 from typing import Any, List
 
-# from edsl.inference_services.InferenceServiceABC import InferenceServiceABC
-from ..language_models import LanguageModel
-
-from ..open_ai_service import OpenAIService
+from ...language_models import LanguageModel
+from .open_ai_service import OpenAIService
 
 
 class DeepSeekService(OpenAIService):
