+++ edsl/inference_services/services/mistral_ai_service.py
@@ -1,11 +1,11 @@
 import os
 from typing import Any, List, Optional
-from .inference_service_abc import InferenceServiceABC
-from ..language_models import LanguageModel
 import asyncio
 from mistralai import Mistral
 
-from ..language_models import LanguageModelBadResponseError
+
+from ..inference_service_abc import InferenceServiceABC
+from ...language_models import LanguageModel
 
 
 class MistralAIService(InferenceServiceABC):
