+++ edsl/inference_services/services/perplexity_service.py
@@ -1,13 +1,9 @@
-import aiohttp
-import json
-import requests
 from typing import Any, List, Optional
-from edsl.inference_services.rate_limits_cache import rate_limits
+from ..rate_limits_cache import rate_limits
 
-# from edsl.inference_services.InferenceServiceABC import InferenceServiceABC
-from ..language_models import LanguageModel
+from ...language_models import LanguageModel
 
-from ..open_ai_service import OpenAIService
+from .open_ai_service import OpenAIService
 
 
 class PerplexityService(OpenAIService):
