+++ edsl/language_models/model.py
@@ -5,12 +5,8 @@ from typing import Optional, TYPE_CHECKING, List, Callable
 from ..utilities import PrettyList
 from ..config import CONFIG
 
-from edsl.inference_services.InferenceServicesCollection import (
-    InferenceServicesCollection,
-)
-from edsl.inference_services.data_structures import AvailableModels
-from edsl.inference_services.InferenceServiceABC import InferenceServiceABC
-from edsl.exceptions.inference_services import InferenceServiceError
+from ..inference_services import (InferenceServicesCollection, 
+                                  AvailableModels, InferenceServiceABC, InferenceServiceError, default)
 
 from ..enums import InferenceServiceLiteral
 
@@ -23,15 +19,12 @@ def get_model_class(
     registry: Optional[InferenceServicesCollection] = None,
     service_name: Optional[InferenceServiceLiteral] = None,
 ):
-    from edsl.inference_services.registry import default
-
     registry = registry or default
     try:
         factory = registry.create_model_factory(model_name, service_name=service_name)
         return factory
     except (InferenceServiceError, Exception) as e:
         return Model._handle_model_error(model_name, e)
-
 
 class Meta(type):
     def __repr__(cls):
@@ -60,8 +53,6 @@ class Model(metaclass=Meta):
     def get_registry(cls) -> InferenceServicesCollection:
         """Get the current registry or initialize with default if None"""
         if cls._registry is None:
-            from edsl.inference_services.registry import default
-
             cls._registry = default
         return cls._registry
 
@@ -289,7 +280,7 @@ class Model(metaclass=Meta):
         works_with_text: Optional[bool] = None,
         works_with_images: Optional[bool] = None,
     ) -> list[dict]:
-        from edsl.coop import Coop
+        from ..coop import Coop
 
         c = Coop()
         working_models = c.fetch_working_models()
